+++ src/core/state/keys.ts
@@ -1,0 +1,67 @@
+export type SecretKey =
+	| "apiKey"
+	| "clineApiKey"
+	| "openRouterApiKey"
+	| "awsAccessKey"
+	| "awsSecretKey"
+	| "awsSessionToken"
+	| "openAiApiKey"
+	| "geminiApiKey"
+	| "openAiNativeApiKey"
+	| "deepSeekApiKey"
+	| "requestyApiKey"
+	| "togetherApiKey"
+	| "qwenApiKey"
+	| "mistralApiKey"
+	| "liteLlmApiKey"
+	| "authNonce"
+	| "asksageApiKey"
+	| "xaiApiKey"
+	| "sambanovaApiKey"
+export type GlobalStateKey =
+	| "apiProvider"
+	| "apiModelId"
+	| "awsRegion"
+	| "awsUseCrossRegionInference"
+	| "awsBedrockUsePromptCache"
+	| "awsBedrockEndpoint"
+	| "awsProfile"
+	| "awsUseProfile"
+	| "vertexProjectId"
+	| "vertexRegion"
+	| "lastShownAnnouncementId"
+	| "customInstructions"
+	| "taskHistory"
+	| "openAiBaseUrl"
+	| "openAiModelId"
+	| "openAiModelInfo"
+	| "ollamaModelId"
+	| "ollamaBaseUrl"
+	| "ollamaApiOptionsCtxNum"
+	| "lmStudioModelId"
+	| "lmStudioBaseUrl"
+	| "anthropicBaseUrl"
+	| "azureApiVersion"
+	| "openRouterModelId"
+	| "openRouterModelInfo"
+	| "openRouterProviderSorting"
+	| "autoApprovalSettings"
+	| "browserSettings"
+	| "chatSettings"
+	| "vsCodeLmModelSelector"
+	| "userInfo"
+	| "previousModeApiProvider"
+	| "previousModeModelId"
+	| "previousModeThinkingBudgetTokens"
+	| "previousModeVsCodeLmModelSelector"
+	| "previousModeModelInfo"
+	| "liteLlmBaseUrl"
+	| "liteLlmModelId"
+	| "qwenApiLine"
+	| "requestyModelId"
+	| "togetherModelId"
+	| "mcpMarketplaceCatalog"
+	| "telemetrySetting"
+	| "asksageApiUrl"
+	| "thinkingBudgetTokens"
+	| "planActSeparateModelsSetting"
