+++ src/exports/index.ts
@@ -1,27 +1,28 @@
 import * as vscode from "vscode"
-import { ClineProvider } from "../core/webview/ClineProvider"
+import { Controller } from "../core/controller"
 import { ClineAPI } from "./cline"
+import { getGlobalState } from "../core/state"
 
-export function createClineAPI(outputChannel: vscode.OutputChannel, sidebarProvider: ClineProvider): ClineAPI {
+export function createClineAPI(outputChannel: vscode.OutputChannel, sidebarController: Controller): ClineAPI {
 	const api: ClineAPI = {
 		setCustomInstructions: async (value: string) => {
-			await sidebarProvider.updateCustomInstructions(value)
+			await sidebarController.updateCustomInstructions(value)
 			outputChannel.appendLine("Custom instructions set")
 		},
 
 		getCustomInstructions: async () => {
-			return (await sidebarProvider.getGlobalState("customInstructions")) as string | undefined
+			return (await getGlobalState(sidebarController.context, "customInstructions")) as string | undefined
 		},
 
 		startNewTask: async (task?: string, images?: string[]) => {
 			outputChannel.appendLine("Starting new task")
-			await sidebarProvider.clearTask()
-			await sidebarProvider.postStateToWebview()
-			await sidebarProvider.postMessageToWebview({
+			await sidebarController.clearTask()
+			await sidebarController.postStateToWebview()
+			await sidebarController.postMessageToWebview({
 				type: "action",
 				action: "chatButtonClicked",
 			})
-			await sidebarProvider.postMessageToWebview({
+			await sidebarController.postMessageToWebview({
 				type: "invoke",
 				invoke: "sendMessage",
 				text: task,
@@ -36,7 +37,7 @@ export function createClineAPI(outputChannel: vscode.OutputChannel, sidebarProvi
 			outputChannel.appendLine(
 				`Sending message: ${message ? `"${message}"` : "undefined"} with ${images?.length || 0} image(s)`,
 			)
-			await sidebarProvider.postMessageToWebview({
+			await sidebarController.postMessageToWebview({
 				type: "invoke",
 				invoke: "sendMessage",
 				text: message,
@@ -46,7 +47,7 @@ export function createClineAPI(outputChannel: vscode.OutputChannel, sidebarProvi
 
 		pressPrimaryButton: async () => {
 			outputChannel.appendLine("Pressing primary button")
-			await sidebarProvider.postMessageToWebview({
+			await sidebarController.postMessageToWebview({
 				type: "invoke",
 				invoke: "primaryButtonClick",
 			})
@@ -54,7 +55,7 @@ export function createClineAPI(outputChannel: vscode.OutputChannel, sidebarProvi
 
 		pressSecondaryButton: async () => {
 			outputChannel.appendLine("Pressing secondary button")
-			await sidebarProvider.postMessageToWebview({
+			await sidebarController.postMessageToWebview({
 				type: "invoke",
 				invoke: "secondaryButtonClick",
 			})
