+++ src/extension.ts
@@ -2,13 +2,13 @@
 // Import the module and reference it with the alias vscode in your code below
 import { setTimeout as setTimeoutPromise } from "node:timers/promises"
 import * as vscode from "vscode"
-import { ClineProvider } from "./core/webview/ClineProvider"
 import { Logger } from "./services/logging/Logger"
 import { createClineAPI } from "./exports"
 import "./utils/path" // necessary to have access to String.prototype.toPosix
 import { DIFF_VIEW_URI_SCHEME } from "./integrations/editor/DiffViewProvider"
 import assert from "node:assert"
 import { telemetryService } from "./services/telemetry/TelemetryService"
+import { WebviewProvider } from "./core/webview"
 
 /*
 Built using https://github.com/microsoft/vscode-webview-ui-toolkit
@@ -30,12 +30,12 @@ export function activate(context: vscode.ExtensionContext) {
 	Logger.initialize(outputChannel)
 	Logger.log("Cline extension activated")
 
-	const sidebarProvider = new ClineProvider(context, outputChannel)
+	const sidebarWebview = new WebviewProvider(context, outputChannel)
 
 	vscode.commands.executeCommand("setContext", "cline.isDevMode", IS_DEV && IS_DEV === "true")
 
 	context.subscriptions.push(
-		vscode.window.registerWebviewViewProvider(ClineProvider.sideBarId, sidebarProvider, {
+		vscode.window.registerWebviewViewProvider(WebviewProvider.sideBarId, sidebarWebview, {
 			webviewOptions: { retainContextWhenHidden: true },
 		}),
 	)
@@ -43,15 +43,15 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		vscode.commands.registerCommand("cline.plusButtonClicked", async () => {
 			Logger.log("Plus button Clicked")
-			const visibleProvider = ClineProvider.getVisibleInstance()
-			if (!visibleProvider) {
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			if (!visibleWebview) {
 				Logger.log("Cannot find any visible Cline instances.")
 				return
 			}
 
-			await visibleProvider.clearTask()
-			await visibleProvider.postStateToWebview()
-			await visibleProvider.postMessageToWebview({
+			await visibleWebview.controller.clearTask()
+			await visibleWebview.controller.postStateToWebview()
+			await visibleWebview.controller.postMessageToWebview({
 				type: "action",
 				action: "chatButtonClicked",
 			})
@@ -60,13 +60,13 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand("cline.mcpButtonClicked", () => {
-			const visibleProvider = ClineProvider.getVisibleInstance()
-			if (!visibleProvider) {
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			if (!visibleWebview) {
 				Logger.log("Cannot find any visible Cline instances.")
 				return
 			}
 
-			visibleProvider.postMessageToWebview({
+			visibleWebview.controller.postMessageToWebview({
 				type: "action",
 				action: "mcpButtonClicked",
 			})
@@ -77,7 +77,7 @@ export function activate(context: vscode.ExtensionContext) {
 		Logger.log("Opening Cline in new tab")
 		// (this example uses webviewProvider activation event which is necessary to deserialize cached webview, but since we use retainContextWhenHidden, we don't need to use that event)
 		// https://github.com/microsoft/vscode-extension-samples/blob/main/webview-sample/src/extension.ts
-		const tabProvider = new ClineProvider(context, outputChannel)
+		const tabWebview = new WebviewProvider(context, outputChannel)
 		//const column = vscode.window.activeTextEditor ? vscode.window.activeTextEditor.viewColumn : undefined
 		const lastCol = Math.max(...vscode.window.visibleTextEditors.map((editor) => editor.viewColumn || 0))
 
@@ -88,7 +88,7 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 		const targetCol = hasVisibleEditors ? Math.max(lastCol + 1, 1) : vscode.ViewColumn.Two
 
-		const panel = vscode.window.createWebviewPanel(ClineProvider.tabPanelId, "Cline", targetCol, {
+		const panel = vscode.window.createWebviewPanel(WebviewProvider.tabPanelId, "Cline", targetCol, {
 			enableScripts: true,
 			retainContextWhenHidden: true,
 			localResourceRoots: [context.extensionUri],
@@ -99,7 +99,7 @@ export function activate(context: vscode.ExtensionContext) {
 			light: vscode.Uri.joinPath(context.extensionUri, "assets", "icons", "robot_panel_light.png"),
 			dark: vscode.Uri.joinPath(context.extensionUri, "assets", "icons", "robot_panel_dark.png"),
 		}
-		tabProvider.resolveWebviewView(panel)
+		tabWebview.resolveWebviewView(panel)
 
 		// Lock the editor group so clicking on files doesn't open them over the panel
 		await setTimeoutPromise(100)
@@ -112,13 +112,13 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		vscode.commands.registerCommand("cline.settingsButtonClicked", () => {
 			//vscode.window.showInformationMessage(message)
-			const visibleClineProvider = ClineProvider.getVisibleInstance()
-			if (!visibleClineProvider) {
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			if (!visibleWebview) {
 				Logger.log("Cannot find any visible Cline instances.")
 				return
 			}
 
-			visibleClineProvider.postMessageToWebview({
+			visibleWebview.controller.postMessageToWebview({
 				type: "action",
 				action: "settingsButtonClicked",
 			})
@@ -127,13 +127,13 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand("cline.historyButtonClicked", () => {
-			const visibleProvider = ClineProvider.getVisibleInstance()
-			if (!visibleProvider) {
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			if (!visibleWebview) {
 				Logger.log("Cannot find any visible Cline instances.")
 				return
 			}
 
-			visibleProvider.postMessageToWebview({
+			visibleWebview.controller.postMessageToWebview({
 				type: "action",
 				action: "historyButtonClicked",
 			})
@@ -142,13 +142,13 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(
 		vscode.commands.registerCommand("cline.accountButtonClicked", () => {
-			const visibleProvider = ClineProvider.getVisibleInstance()
-			if (!visibleProvider) {
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			if (!visibleWebview) {
 				Logger.log("Cannot find any visible Cline instances.")
 				return
 			}
 
-			visibleProvider.postMessageToWebview({
+			visibleWebview.controller.postMessageToWebview({
 				type: "action",
 				action: "accountButtonClicked",
 			})
@@ -179,15 +179,15 @@ export function activate(context: vscode.ExtensionContext) {
 
 		const path = uri.path
 		const query = new URLSearchParams(uri.query.replace(/\+/g, "%2B"))
-		const visibleProvider = ClineProvider.getVisibleInstance()
-		if (!visibleProvider) {
+		const visibleWebview = WebviewProvider.getVisibleInstance()
+		if (!visibleWebview) {
 			return
 		}
 		switch (path) {
 			case "/openrouter": {
 				const code = query.get("code")
 				if (code) {
-					await visibleProvider.handleOpenRouterCallback(code)
+					await visibleWebview?.controller.handleOpenRouterCallback(code)
 				}
 				break
 			}
@@ -203,13 +203,13 @@ export function activate(context: vscode.ExtensionContext) {
 				})
 
 				// Validate state parameter
-				if (!(await visibleProvider.validateAuthState(state))) {
+				if (!(await visibleWebview?.controller.validateAuthState(state))) {
 					vscode.window.showErrorMessage("Invalid auth state")
 					return
 				}
 
 				if (token && apiKey) {
-					await visibleProvider.handleAuthCallback(token, apiKey)
+					await visibleWebview?.controller.handleAuthCallback(token, apiKey)
 				}
 				break
 			}
@@ -224,7 +224,7 @@ export function activate(context: vscode.ExtensionContext) {
 		// Use dynamic import to avoid loading the module in production
 		import("./dev/commands/tasks")
 			.then((module) => {
-				const devTaskCommands = module.registerTaskCommands(context, sidebarProvider)
+				const devTaskCommands = module.registerTaskCommands(context, sidebarWebview.controller)
 				context.subscriptions.push(...devTaskCommands)
 				Logger.log("Cline dev task commands registered")
 			})
@@ -253,8 +253,8 @@ export function activate(context: vscode.ExtensionContext) {
 			const filePath = editor.document.uri.fsPath
 			const languageId = editor.document.languageId
 
-			// Send to sidebar provider
-			await sidebarProvider.addSelectedCodeToChat(
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			await visibleWebview?.controller.addSelectedCodeToChat(
 				selectedText,
 				filePath,
 				languageId,
@@ -303,7 +303,8 @@ export function activate(context: vscode.ExtensionContext) {
 				*/
 
 				// Send to sidebar provider
-				await sidebarProvider.addSelectedTerminalOutputToChat(terminalContents, terminal.name)
+				const visibleWebview = WebviewProvider.getVisibleInstance()
+				await visibleWebview?.controller.addSelectedTerminalOutputToChat(terminalContents, terminal.name)
 			} catch (error) {
 				// Ensure clipboard is restored even if an error occurs
 				await vscode.env.clipboard.writeText(tempCopyBuffer)
@@ -374,11 +375,12 @@ export function activate(context: vscode.ExtensionContext) {
 			const languageId = editor.document.languageId
 
 			// Send to sidebar provider with diagnostics
-			await sidebarProvider.fixWithCline(selectedText, filePath, languageId, diagnostics)
+			const visibleWebview = WebviewProvider.getVisibleInstance()
+			await visibleWebview?.controller.fixWithCline(selectedText, filePath, languageId, diagnostics)
 		}),
 	)
 
-	return createClineAPI(outputChannel, sidebarProvider)
+	return createClineAPI(outputChannel, sidebarWebview.controller)
 }
 
 // This method is called when your extension is deactivated
