+++ src/integrations/checkpoints/CheckpointTracker-old.ts
@@ -3,7 +3,7 @@ import os from "os"
 import * as path from "path"
 import simpleGit, { SimpleGit } from "simple-git"
 import * as vscode from "vscode"
-import { ClineProvider } from "../../core/webview/ClineProvider"
+import { Controller as ClineProvider } from "../../core/controller"
 import { fileExistsAtPath } from "../../utils/fs"
 import { globby } from "globby"
